Refactor database index creation
This commit refactors the code in the `create_indexes` method of the `Database` class in `database.py`. It adds the `background=True` parameter to the `create_index` calls, allowing the indexes to be created in the background. This improves performance by not blocking other operations while the indexes are being created.